Add `StringSyntax` to the `RouteAttribute`
I noticed when using the `[Route("/orders/{orderId:int}")]` attribute, as an example, doesn't display the same as when using the `EndpointRouteBuilderExtensions` APIs for `Map*` that accept a template/pattern. I think it would be great if we could also highlight routes when using the attribute. This PR adds the `StringSyntax` to decorate the `template` parameter to be highlighted like the method APIs.