Add hpack decoding test (#57221)
* Add hpack decoding test

There's a note in section 4.4 of the spec that an incoming header field can reference a dynamic table entry (i.e. for its name) that will be evicted as the incoming header field is inserted into the dynamic table.  Add a test validating that we do something sensible in this case.

* Reuse existing large string literal