Unquarantine ServerReset_BeforeRequestBody_ClientBodyThrows
Fixes #46717

There isn't an obvious race in the test and the component that timed out appears to be from runtime (`HttpClient` and/or `HttpContent`).  Since the test hasn't failed in 30 days and runtime has likely changed in the 18 months since the issue was filed, I think we should unquarantine the test.