Fully Qualify Type Names When Resolving in RDG
These two places correspond to resolution of a service from DI when
attributed with `FromServicesAttribute` or with
`FromKeyedServicesAttribute`. When the type name of the service is not
fully qualified, the generator can generate code which fails to compile.

For example, given a service `ExampleService` in namespace `Http`, the
generator can generate code to resolve an instance of such a service
like this:

```c#
var e_local = httpContext.RequestServices.GetRequiredKeyedService<Http.ExampleService>("example");
```

…but `Microsoft.AspNetCore.Http` is a nearer match for the reference to
"Http", so the compiler will look for a type
`Microsoft.AspNetCore.Http.ExampleService` (which doesn't exist – at
least, not at time of writing), and compilation will fail.

Fixes: #58633