Improve performance of UriHelper.GetDisplayUrl (#55611)
* Optimized string concatenation with `string.Create`

Replaced `StringBuilder` concatenation with the more efficient `string.Create` method for creating new strings by concatenating `scheme`, `host`, `pathBase`, `path`, and `queryString`. This method reduces overhead by avoiding multiple string concatenations and allocating the correct buffer size for the new string. The `CopyTo` method is used in a callback function to copy each string to the new string, slicing the buffer to remove the copied part. The `SchemeDelimiter` is always copied to the new string, regardless of its length. This change enhances the performance of the code.

* changed to string.Concat(params ReadOnlySpan<string?>)

* Update src/Http/Http.Extensions/src/UriHelper.cs

---------

Co-authored-by: Stephen Halter <halter73@gmail.com>