Update CODEOWNERS to include doc writers and reduce duplication
This PR adds @Rick-Anderson and @tdykstra as CODEOWNERS for `/**/PublicAPI.*Shipped.txt`, so they get notified for any PR that adds new API so they can review the doc comments.

Alternatively, they could join @aspnet-api-review, but that may be used for more than just PRs.

I also took the opportunity to remove some redundancy with the assumption that the [last matched pattern wins](https://graphite.dev/guides/in-depth-guide-github-codeowners#patterns-and-precedence-in-the-codeowners-file). If we find that's not the case, we can revert that part later and add @Rick-Anderson and @tdykstra to all the relevant lines.