Fixed the issue. Will be adding the tests as a separate commit. (#60096)
# Include typed result metadata in the action descriptors for MVC Controller actions

## Description

The endpoint metadata wasn't being captured in ActionModel during the ApplicationModel creation. This is addressed by the change in the DefaultApplicationModelProvider class. The metadata is now being extracted based on the return type of the action and then the selectors associated with the action are populated with that metadata, as that's what is being used later on for populating the ActionDescriptor's EndpointMetadata.

Fixes #44988