Merged PR 47890: [2.3] Prevent RefreshSignInAsync if it is called with wrong user
Since this is a patch, instead of throwing an exception in cases where we wouldn't before like we do in the PR to main, we instead log an error and fail to refresh the cookie if RefreshSignInAsync is called with a TUser that does not have the same user ID as the currently authenticated user. While this does not make it *as* obvious to developers that something has gone wrong, error logs are still pretty visible, and stale cookies are something web developers have to account for regardless.