Use Microsoft.NETCore.App.Ref version in place of legacy MicrosoftNETCoreAppPackageVersion
https://github.com/dotnet/dotnet/blob/main/repo-projects/Directory.Build.props#L247-L251

The reason for this block is that repos tend to use arch specific version properties as proxy version numbers. So aspnetcore will set the runtime pack version based on the MicrosoftNETCoreAppVersion, which it typically sets based on the x64 runtime pack version. But since that's not produced in an x86 build, we jump some through VMR hoops to get this right.
The stabilized build exposes a few of these cases where we incorrectly use a few properties, which is quite fixable, but overall this block caught my attention and I think we should remove it.
aspnetcore seems to be the biggest user of MicrosoftNETCoreAppVersion, set to the x64 runtime pack version. I was thinking to just changing all the uses to the ref pack version, since the runtime pack and ref pack versions are always aligned. We can then get rid of this block altogether, which would be great