Optimize prefixed parameters when assembling model name.
Added an method to allow optimize model names where the model name is the same
as one of its properties when binding the model name. This should only
be used when this is the desired behavior. Tests have been added to
assert this behavior when passing parameters whose model name is the
same as the property to be bound.

This was introduced to allow circumventing the bug where the parameter
name is the same as the name of one of its properties, causing the
model to trick the PrefixContainer into considering that parameter
as prefixed (model.property), even when it is not.

```C#
class SomeClass
{
  public string Parameter { get; set; } = "";
}
```

and later we define the model name as parameter (e.g in a ActionMethod for example)

```C#
public IActionResult SomeEndpoint([FromQuery] SomeClass parameter) { /* elided */ }
```

internally, when we pass the query "?parameter=somecoolparametervalue", it will classify
it as prefixed (parameter.Parameter instead of only parameter), because of the internal
logic of PrefixContainer. However, it is not prefixed, so later on when it try to bind the
key (the keys from the query, in this case parameter) with the internal keys (wrong classified
as parameter.Parameter) the ModelBinder would not find an valid candidate for matching with the
key.

Now the optimization option allows to treat the modelName and the propertyName with only one if
both are equal.