Fix post actions for Razor Class Library template
The Razor Class Library template has an option, `SupportsPagesAndViews`
that causes it to produce one of two sets of files. If true, then
`Areas/MyFeature/Pages/Page1.cshtml` is created. If false, then
`Component1.razor` is created.

The issue with the previous template is that the set of primary outputs
and the corresponding `openInEditor` post-actions were not taking this
option into account, and were always trying to open the `Page1.cshtml`
file, even if it didn't exist on disk.

This led to the bug described in https://github.com/microsoft/vscode-dotnettools/issues/2489

C# Dev Kit has been patched to not show warnings when a template
specifies that non-existent files should be opened.

With the change here, users creating this template will now see the
correct file opened. This applies both to VS and CDK.