Remove retrials of fixed tests
Updated according to https://github.com/dotnet/aspnetcore-internal/issues/4769#issuecomment-3642606696.

For these 2 removals we know what fixed it.
`CertificateChangedOnDisk` might be lucky to not pass for the last 90 days. Leaving on the list.
`CheckInvalidHostingModelParameter` - due to its dependence on deployment, HTTP, event logging I am not 100% sure if we should remove it. It's up to discussion or can be done later when another 90 days of all-green period will be confirmed.