Fix exit code handling in runtests.cmd (#64820)
We were inadvertently setting exit code = 0 even if a test failed.

This is because the way the `errorlevel` _command_ works is that it checks whether the last command exit code was **greater than or equal** to the argument.

So checking `errorlevel 0` in the if reports true even when we have exit code 1 and then we negate that and don't enter the if block.

We don't actually need this if check at all since we just want to print the value so we can just capture the `%errorlevel%` _variable_.